Extracted code that creates server service into separate file.
